requirements.txt
Flask==3.1.3
Flask-CORS==6.0.2
Flask-SocketIO==5.6.0
opencv-python-headless==4.13.0.92
ultralytics==8.4.14
numpy==2.4.2
python-dotenv==1.2.1
requests==2.32.5
pillow==12.1.1
werkzeug==3.1.5
librosa==0.11.0
tensorflow==2.20.0
soundfile==0.13.1